tor_rss_reader/requirements-build.txt
# Runtime-Abhängigkeiten
feedparser==6.0.10
requests==2.31.0
PySocks==1.7.1
urllib3==2.0.7

# Build-Abhängigkeiten
pyinstaller==6.3.0
pillow==10.1.0
